Architecture fix
Fixed architecture according onion architecture

+++ ItechArtLabPetsitters.Infrastructure/Repository/EFRepository/ServiceEFRepository.cs
@@ -31,19 +31,21 @@ namespace ItechArtLabPetsitters.Infrastructure.Repository.EFRepository
         {
             return await _dbContext.Services.FirstAsync(p=>p.ID==ID);
         }
-        public async Task <ActionResult> AddServiceAsync(ServiceCreationModel model) 
+        public async Task AddServiceAsync(ServiceCreationModel model) 
         {
             Service AddedService = mapper.Map<ServiceCreationModel, Service>(model);
             _dbContext.Services.Add(AddedService);
             await _dbContext.SaveChangesAsync();
-            return new OkResult();
         }
-        public async Task <ActionResult> DeleteServiceAsync(long id) 
+        public async Task DeleteServiceAsync(long id) 
         {
-            Service deletedService = await _dbContext.Services.FirstAsync(p => p.ID == id); 
-            _dbContext.Services.Remove(deletedService);
-            await _dbContext.SaveChangesAsync();
-            return new OkResult();
+            Service deletedService = await _dbContext.Services.FirstAsync(p => p.ID == id);
+            if (deletedService != null) 
+            {
+                _dbContext.Services.Remove(deletedService);
+                await _dbContext.SaveChangesAsync();
+            }
+            
         }
     }
 }
